move suggestion module tests
